Fix PTP_SYS_OFFSET_EXTENDED returning non-monotonous timestamps
For some reason the driver may be unable to estimate PCI timing.
In that situation (which happens 100% of the time on the various hardware I tried)
the ts_window_adjust would be assinged to (u32)((U64_MAX >> 5) * 3) = 4294967293

because of this PTP_SYS_OFFSET_EXTENDED's post_ts is substracted 4.3s, making it smaller than pre_ts.